src/entrega_2/backend/requirements.txt
# FastAPI & Server
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary
alembic

# Authentication
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# CORS & Utils
pydantic[email-validator]
python-dotenv

# Export
openpyxl
